+++ cedar-policy-core/src/tpe/response.rs
@@ -59,8 +59,8 @@ impl ResidualPolicy {
     }
 
     /// Get the [`PolicyID`]
-    pub fn get_policy_id(&self) -> PolicyID {
-        self.policy.id().clone()
+    pub fn get_policy_id(&self) -> &PolicyID {
+        self.policy.id()
     }
 
     /// All literal uids referenced by this residual
@@ -87,16 +87,20 @@ impl From<ResidualPolicy> for Policy {
 pub struct Response<'a> {
     decision: Option<Decision>,
     residuals: HashMap<PolicyID, ResidualPolicy>,
-    // All of the [`Effect::Permit`] policies that were satisfied
-    satisfied_permits: HashSet<PolicyID>,
-    // All of the [`Effect::Permit`] policies that were not satisfied
+    // All of the [`Effect::Permit`] policies that were true
+    true_permits: HashSet<PolicyID>,
+    // All of the [`Effect::Permit`] policies that were false
     false_permits: HashSet<PolicyID>,
+    // All of the [`Effect::Permit`] policies that errored
+    error_permits: HashSet<PolicyID>,
     // All of the [`Effect::Permit`] policies that evaluated to a residual
     residual_permits: HashSet<PolicyID>,
-    // All of the [`Effect::Forbid`] policies that were satisfied
-    satisfied_forbids: HashSet<PolicyID>,
-    // All of the [`Effect::Forbid`] policies that were not satisfied
+    // All of the [`Effect::Forbid`] policies that were true
+    true_forbids: HashSet<PolicyID>,
+    // All of the [`Effect::Forbid`] policies that were false
     false_forbids: HashSet<PolicyID>,
+    // All of the [`Effect::Forbid`] policies that errored
+    error_forbids: HashSet<PolicyID>,
     // All of the [`Effect::Forbid`] policies that evaluated to a residual
     residual_forbids: HashSet<PolicyID>,
     // request used for this partial evaluation
@@ -117,11 +121,13 @@ impl<'a> Response<'a> {
         schema: &'a ValidatorSchema,
     ) -> Self {
         let mut residual_map = HashMap::new();
-        let mut satisfied_permits = HashSet::new();
+        let mut true_permits = HashSet::new();
         let mut false_permits = HashSet::new();
+        let mut error_permits = HashSet::new();
         let mut residual_permits = HashSet::new();
-        let mut satisfied_forbids = HashSet::new();
+        let mut true_forbids = HashSet::new();
         let mut false_forbids = HashSet::new();
+        let mut error_forbids = HashSet::new();
         let mut residual_forbids = HashSet::new();
         for rp in residuals {
             let r = rp.get_residual();
@@ -130,28 +136,32 @@ impl<'a> Response<'a> {
             match rp.get_effect() {
                 Effect::Forbid => {
                     if r.is_true() {
-                        satisfied_forbids.insert(id);
-                    } else if r.is_false() || r.is_error() {
-                        false_forbids.insert(id);
+                        true_forbids.insert(id.clone());
+                    } else if r.is_false() {
+                        false_forbids.insert(id.clone());
+                    } else if r.is_error() {
+                        error_forbids.insert(id.clone());
                     } else {
-                        residual_forbids.insert(id);
+                        residual_forbids.insert(id.clone());
                     }
                 }
                 Effect::Permit => {
                     if r.is_true() {
-                        satisfied_permits.insert(id);
-                    } else if r.is_false() || r.is_error() {
-                        false_permits.insert(id);
+                        true_permits.insert(id.clone());
+                    } else if r.is_false() {
+                        false_permits.insert(id.clone());
+                    } else if r.is_error() {
+                        error_permits.insert(id.clone());
                     } else {
-                        residual_permits.insert(id);
+                        residual_permits.insert(id.clone());
                     }
                 }
             }
         }
 
         let decision = match (
-            !satisfied_forbids.is_empty(),
-            !satisfied_permits.is_empty(),
+            !true_forbids.is_empty(),
+            !true_permits.is_empty(),
             !residual_permits.is_empty(),
             !residual_forbids.is_empty(),
         ) {
@@ -170,11 +180,13 @@ impl<'a> Response<'a> {
         Self {
             decision,
             residuals: residual_map,
-            satisfied_permits,
+            true_permits,
             false_permits,
+            error_permits,
             residual_permits,
-            satisfied_forbids,
+            true_forbids,
             false_forbids,
+            error_forbids,
             residual_forbids,
             request,
             entities,
@@ -188,7 +200,7 @@ impl<'a> Response<'a> {
             clippy::unwrap_used,
             reason = "we know that the policy ids are in the residuals map"
         )]
-        self.satisfied_permits
+        self.true_permits
             .iter()
             .map(|id| self.residuals.get(id).unwrap())
     }
@@ -199,7 +211,7 @@ impl<'a> Response<'a> {
             clippy::unwrap_used,
             reason = "we know that the policy ids are in the residuals map"
         )]
-        self.satisfied_forbids
+        self.true_forbids
             .iter()
             .map(|id| self.residuals.get(id).unwrap())
     }
@@ -215,6 +227,17 @@ impl<'a> Response<'a> {
             .map(|id| self.residuals.get(id).unwrap())
     }
 
+    /// Get trivially erroring permit residual policies
+    pub fn error_permits(&self) -> impl Iterator<Item = &ResidualPolicy> {
+        #[expect(
+            clippy::unwrap_used,
+            reason = "we know that the policy ids are in the residuals map"
+        )]
+        self.error_permits
+            .iter()
+            .map(|id| self.residuals.get(id).unwrap())
+    }
+
     /// Get trivially false forbid residual policies
     pub fn false_forbids(&self) -> impl Iterator<Item = &ResidualPolicy> {
         #[expect(
@@ -222,6 +245,17 @@ impl<'a> Response<'a> {
             reason = "we know that the policy ids are in the residuals map"
         )]
         self.false_forbids
+            .iter()
+            .map(|id| self.residuals.get(id).unwrap())
+    }
+
+    /// Get trivially erroring forbid residual policies
+    pub fn error_forbids(&self) -> impl Iterator<Item = &ResidualPolicy> {
+        #[expect(
+            clippy::unwrap_used,
+            reason = "we know that the policy ids are in the residuals map"
+        )]
+        self.error_forbids
             .iter()
             .map(|id| self.residuals.get(id).unwrap())
     }
@@ -256,6 +290,14 @@ impl<'a> Response<'a> {
     /// Attempt to get the authorization decision
     pub fn decision(&self) -> Option<Decision> {
         self.decision
+    }
+
+    /// Get the determining policies for the authorization decision
+    pub fn reason(&self) -> Option<impl Iterator<Item = &PolicyID>> {
+        match self.decision? {
+            Decision::Allow => Some(self.true_permits.iter()),
+            Decision::Deny => Some(self.true_forbids.iter()),
+        }
     }
 
     /// Perform reauthorization